Don't call Integer.intValue(), just let the compiler unbox.
git-svn-id: https://svn.code.sf.net/p/vassalengine/svn/VASSAL-src/trunk@9390 67b53d14-2c14-4ace-a08f-0dab2b34000c

+++ src/VASSAL/chat/ServerAddressBook.java
@@ -743,7 +743,7 @@ public class ServerAddressBook {
             Resources.getString("ServerAddressBook.edit_server_configuration"), //$NON-NLS-1$
             config.getControls(), JOptionPane.PLAIN_MESSAGE, null, JOptionPane.OK_CANCEL_OPTION,
             null, null, null, null);
-        if (result != null && result.intValue() == 0) {
+        if (result != null && result == 0) {
           if (enabled) {
             setProperties(config.getProperties());
             saveAddressBook();
