+++ src/VASSAL/chat/jabber/JabberRoom.java
@@ -360,8 +360,8 @@ public class JabberRoom extends SimpleRoom implements LockableRoom {
 
   public static Properties configureNewRoom() {
     final JabberRoomConfig config = new JabberRoomConfig();
-    final Integer result = ((Integer) Dialogs.showDialog(null, Resources.getString("Chat.create_new_room"), config, JOptionPane.PLAIN_MESSAGE, null, JOptionPane.OK_CANCEL_OPTION, null, null, null, null)).intValue(); //$NON-NLS-1$
-    if (result != null && result.intValue() == 0) {
+    final Integer result = (Integer) Dialogs.showDialog(null, Resources.getString("Chat.create_new_room"), config, JOptionPane.PLAIN_MESSAGE, null, JOptionPane.OK_CANCEL_OPTION, null, null, null, null); //$NON-NLS-1$
+    if (result != null && result == 0) {
       return config.getProperties();
     }
     return null;
